many adjustments and fixes for multiplayer
heal formula now goes through element/damage calculation. it is no longer hard-coded to give HP to the player that cast it: it generates a large hitbox, and determines its heal/damage value through the elements of the colliding boxes. heal spells are of the Verve element (9). the verve element applies a -1 modifier to the formula's potency, unless the target's affinity is of the Entropy element (10). in which case the modifier is set to 1 (normal damage). before displaying the damage popup, we check if the damageTaken variable is less than 0. if damageTaken is negative, that means it is Healing; the damage popup displays the value as positive and cyan-coloured, indicating healing.
the biggest issue this has caused, which i should have anticipated, is that since healing spells/buffs look for monitoring Player Hurtboxes, they will not get hit with the heal or the buff if they have invincibility from just getting hit, as invincibility disables Hurtbox monitoring briefly. so i've added another collision layer called FormulaHurtbox, to replace the player's Collectbox (which used to be used for absorbing hearts & coins).

i have also re-programmed the controls. i've had to more discretely separate the different movement/actions buttons for the different players. what's more of a pain is UI screens, and determining which inputs to listen for, as controls ALWAYS listen for inputs preceded with 'ui_', which is a fixture of the engine, unfortunately. in order to prevent player 1 from controlling player 2's formula target screen and vice versa, the target screen now grabs the name of the player who isntantiated it and matches it with the names of the Inputs they should use. i have yet to do this for the LevelUpScreen, so both players can still move focus there.

there have been a lot of system updates this commit so will need some testing as it's possible some errors remain.