created GameManager.save_game() method
the function can be called by going to "save" in the Config menu of the pause screen.
right now, 2 nodes are considered "persistent": 1. the PlayerStats, which saves data about the Player's current stats and experience/level; and the Player, to which relevant resources are attached (inventory, pouch, formulabook).
the save_data() function has successfully created a file which stores this data in JSON format.
next is to write the load function, then i assume a bunch of experimenting to figure out how i want to handle saving/loading/dying/continuing etc.
since player resources are going to be saved to the SAVE file, i'm not sure i need to save/load references to those resources on player initialization anymore. will have to experiment.