need to work on how scenes build themselves before continuing with Load
currently, in each map scene, there is a Player instance as a child of YSort. this means that whenever a map scene is loaded, this particular Player instance is also loaded automatically. this may not be desireable.
the way we worked around this was by saving references to relevant player resources (pouch, formulabook) in the Exit scene and passing them in a dictionary as the argument for the Global scene changer, which applies those resources to the Player when respawned.
now, i am thinking of having the map scene itself create the player via code, rather than have the Player already there. this is likely to cause issues. one issue is that the player needs a position. a "player_spawn_pos" can be declared in each map's script, signifying a default spawn position. this would have to be overridden when zoning from somewhere else, though.