fixing some errors
when a player is added to the SceneTree, it checks Global.get_attribute("location"), which is set with the position the player should spawn at when transitioning to a new zone. if there is no value here, grabs the value from get_tree().get_root().get_node("/root/World/Map").player_spawn_pos.