saving/loading stats works... kind of
but this is still convoluted as all hell. the PlayerStats scene is always available. it gets instantiated ONCE when the program runs, and the same instance remains until the program closes. when the program runs, the default stats - as in, the ones that are hard-coded - are applied. if a game is loaded, PlayerStats gets the values written in the savegame.save file and applies them to the various stats. the problem is on selecting a New Game AFTER selecting Continue, it still uses the old PlayerStats. so i guess a "default_stats" function needs to get run on New Game that resets every stat to its initial value.