saving and loading... working?
the save load system is a zombie pieced together with remnants of old code and badly written new code but somehow, SOMEHOW, it appears to be working as intended. i think.
on new game, the chapter display shows.
on continue, it doesn't.
saving right now saves which map you are on (each map now needs a "default" spawn position), as well as certain PlayerStats (the main ones & health & stamina), and PlayerLog (for chapter_display).
when you save, it writes to file the PlayerLog and PlayerStats into a save dictionary, and the inventory/pouch/formulabook to 3 separate resources.
when the game starts, inventory resources are created right away for the GameManager. the GameManager then checks if the resource files listed above exist on disk, and if they do, it loads them - if they don't, they are created empty.
going to start->config->save saves everything to those files. pressing Continue on the main menu parses through the savegame.save file, finds the map_filename associated with the dict, as well as the appropriate values for PlayerStats/PlayerLog.
note that currently pressing ESC quits the game instantly and goes to the main menu - it does not pause it, it abandons the scene. SPACEBAR pauses the game.
the save system has not yet been implemented for multiplayer. player2 still uses the old methods, so it still works, but his values will not get saved.
FormulaStats also do not persist yet.
another thing i had to do since pressing the alchemy or pouch buttons on the pause screen while having no formulas/ingredients froze the game was disable the buttons if their size == 0.