fixing file issues
i need to have a way to differentiate saving/loading player 1/player 2 stats. perhaps an identifier in the stats scenes themselves. i would like to convert the PlayerStats scene into a resource that can be extended for multiple players as opposed to having 2 discrete instances of the scene autoloaded.
i have disabled all save functionality for Player 2 as player 2's stats would overwrite player 1's when loading due to the nature of how the data is written. when loading, PlayerStats should only look for data previously saved by PlayerStats, and the same for Player2Stats.