trying to clean up the save/load system
pouch/formulabook exist as resources in-game, thus they are saved/loaded as resources discretely. it may be possible to save these resources inside the main .save file, but this involves converting those resources to JSON format, saving them as text, then being able to parse through the save file and find the saved resources as text, parse them, and load them as resources. i don't exactly know how to do that, so for now, the "save" file and the resources associated with the player saved will remain separate.
i have tried my best to streamline the save/load process as much as possible. therefore, the only time we save to or look at the disk is when Save is selected from the pause menu, or when Continue is selected from the main menu. if we load a scene from the editor directly, we do not load resources from the disk: we create new ones, assign default parameters, and save them as a variable in the GameManager... but NOT to disk UNLESS Save is selected.
this is basically for convenience while game designing without necessitating a debug room yet. we can load a scene to test it directly, and get assigned automatically whatever formulas/ingredient amounts/PlayerStats we like. if New Game is selected, the default parameters are assigned: no formulas, no ingredients, and 4 of each stat. if we load a scene directly, we can give our player more stats to facilitate easier testing.
for now, i have completely removed adding in a Player 2. the reason why is i would like to duplicate the AutoLoad Singleton PlayerStats during multiplayer, as opposed to simple having 2 separate singletons, PlayerStats and Player2Stats. currently the only different between the two is in the save() function, where in Player2Stats we define a dict called player2_status containing all of our stats, and assign it to the value of the key "player2_stats" in the save_dict that gets returned when we save. unless that's actually the right way to do it.
maybe another way to do it is to have a generic PlayerStats class from which we extend our stats scene for each player, then each extension can have their own save method. i guess i'll see if this is possible.