save/load fixes for multiplayer
handling saving/loading and drop-in/drop-out multiplayer is complicated for a number of reasons.
1. if the game is played with 1 player only, we don't save resources for player 2. when we load, we check for player 2 resources on disk. if none are found, we skip loading player2.
2. when we save, we check for player2. if there is a player2 active, we save their resources to disk. if there isn't a player2 active BUT WE HAVE DATA FOR PLAYER 2 IN OUR CURRENT SESSION (GameManager.player2_data), we save THAT data to disk instead.

this system will obviously need work when we incorporate a save/progression system into the game.

i am also considering a complete re-work of how PlayerStats are handled. right now, each Player has their own, discrete PlayerStats scene. i think it makes more sense to extend our individual PlayerStats scenes from a single one, and changing only what we need to in those extensions (namely, the save function, since everything else is identical).
the reason i haven't gone ahead and done this yet is that it would require redefining every single reference to PlayerStats in the game right now. i may save that for tomorrow because either it will instantly work as intended, or it will cause so many bugs that it might take an entire day or more to fix.
so as of right now, the game still references the original PlayerStats script, but the Node is no longer in the Persist group so Player1's stats don't get saved - well, technically they do, but none of the game code has been updated to reference the new stats node, so when we level up, we are still updating the OLD PlayerStats node, which is the one that doesn't get saved.
if this works, similar logic can be applied to FormulaStats, and each Player can have their own spell levels/progression.
then, the same idea can be applied to the Player scenes themselves. instead of having 2 discrete scenes for Players, they can both extend from the same scene and have their extensions modify player-specifc parameters (like controls and stat references).
also, by looking through the Enemy script code, i noticed that there are hard-coded references to PlayerStats as opposed to looking for the PlayerStats of the owner of the hitbox that entered the enemy's hurtbox. essentially, what this means (i think) is that parts of the Enemy script ONLY reference Player1's stats.
also fixed a bug where missed attacks against Player2 would still cause status effects.