PlayerStats system overhaul for multiplayer
after a bit of bugfixing, so far everything seems to be working, surprisingly! i also cleaned up some obsolete files, as everything from this point on should take into account multiple players.
the next things to do are to instantiate multiple instances of FormulaStats for each Player, just like we've done with PlayerStats.
then, if i feel like it, do the same thing with the Players themselves: have them all extend from a Player class, and assign unique parameters to each extension, ie. controls and stats... it honestly shouldn't be too complicated since it's basically doing the same thing i've just done with PlayerStats, but it's daunting because the Player script is the densest, most integral part of the entire codebase right now, so i'm anticipating trouble.