deleted old GUI/updated references
was annoyed that everything referenced "GUI2" so i got rid of the OG GUI and replaced it with the new one that incorporates multiplayer.
maps created with the old GUI as a child of World will become broken. i'm thinking of having the World scene, when created, generate its own instance of the GUI through code rather than necessitate each map scene to already have it in the SceneTree. i don't know if that will cause problems or not.