re-wrote Examine code
the system that kept track of what had been examined in order to determine whether to display the ?! icon has been changed. before, each examinable object had their own global variable (in PlayerLog) that remembered whether the object in question had been examined or not. this also determined whether to spawn the object as examined or not.
now, examined objects simply have the name of the object appended to an array, PlayerLog.examined_list. when the object spawns, we check if its name is in the array. if it is, it's been examined. if it's not, it hasn't been.
now that i've done this, it will become easier to further re-factor the examine system. instead of checking in each object's _ready() function if the name of the object is in the examined_list, perhaps every examinable object can be extended from an ExaminableObject class. perhaps that might overcomplicate what we're doing, since there will be many types of objects that are examinable: things in the environment, interactable objects, collectible items, and of course NPCs and other bodies. it also might overcomplicate re-doing the Enemy system as i've wanted to do. currently, common enemy functions are stored in a global Enemy script. it might be better practice if instead of having each enemy reference the same global script, each enemy is EXTENDED from a base Enemy class, which would include all the functions currently in the Enemy script as well as common parameters between enemies. that will be a difficult undertaking as it will required a LOT of refactoring.
also fixed crash on examining formula pickups.
the PlayerLog save data is saved in the main save file along with players' stats and formula data. it is not linked with a particular character: there will only ever be one instance of PlayerLog.