working on save functionality
FormulaData gets reset on new game, and is saved properly for each player.
interactable objects that are toggled have their status saved in the player log.
i figure that each area with toggleable objects with have a dictionary with a key referencing the status of each of those objects. this dictionary will get saved/loaded when the game is saved/loaded. the status of the object is also checked when the object is instantiated: if the PlayerLog lists the object as on - or, more specifically, its key in the area's dictionary has a value of 'true' - it is spawned as on. eg., in the player Home, the state of the lights are saved, as well as the fridge being open, sink running, etc.

i am still encountering unpredictable behaviour regarding saving/loading and multiplayer. every time i think i've solved it, a different issue occurs. once again, i THINK i have solved it. the problem once again is sometimes both players share the same pouch/formulabook resource. meaning if player 1 picks up a formula, both player 1 AND 2 have access to it, or another example, if player 1 uses a formula, it subtracts ingredients from BOTH pouches - in reality, i guess there is only ONE pouch/formulabook resources shared between both players, but i am not sure what makes this so.