cleaning up code
gradually getting rid of the old status "system", and beginning to think about how to refactor the Player and Player2 scenes into a single scene from which we can instance our players.
when we instance different players from a single scene, we will need to provide the parameters that differentiate between the players. for the players, these things will be references to stats and formula data, as well as which inputs to listen for, ie. controls.
i guess next i will just manually go through the scenes and figure out what needs to be changed/done.