working on player2 instantiation methods
the confusing part about making sure this all works is taking into account the multiple contexts in which player2 can be instantiated. one is at the beginning of a scene, if multiplayer is on. player2 is loaded automatically from the world scene, which checks GameManager.multiplayer_2 == true. the other is when player2 is toggled.
maybe it would be good if all of this stuff took place in the GameManager, as the flow of code when players are instantiated jumps around quite a bit.
that aside... things are working! it's messy but i'm glad i was able to instantiate each player from a single scene. will probably take a bit of cleaning up to get right but this will save time in the future, especially if more players are added.