improved interaction/pickup
before, disabled the interactObject (the thing a player would interact with if the "attack" button was pressed) whenever an interactable area was detected leaving the player's interaction zone. now, disables the interactionzone only if the interactObject matches the exiting area's parent. this fixes the issue where if 2 objects were in the player's interaction zone, and 1 leaves, the other would ALSO become un-interactable.