various fixes
"dying" now removes all buffs. previously, Regen ticks would reset the "dying" state. this is no longer the case.
to accomplish this, i've added a new function to the StatusHandler, remove_buffs(body). this will remove all statuses with the "buff" variable from the body entered as an argument.
also, multiplayer can only be toggled while player2 is in the IDLE/MOVE (0) state. this is because certain time-based actions need a reference to the player that started them when they timeout - if the player is toggled off before this timeout occurs, we get an error. for example, when we cast a formula, the ingredients are only decremented after the cast timer. if the player casting is deleted before this cast timer times out, we attempt to decrement ingredients from a player that no longer exists, resulting in an error. player 2 can no longer quit the game while not in the MOVE state (this includes the various ATTACK states, the PICKUP state, ACTION state, STUN state, etc.)