removed continue function
pressing start now goes back to the main menu instead of reviving on the spot. the dying sequence and its functionality was not designed for multiplayer so it was becoming a pain to constantly patch it. a better way would be to redesign the death mechanic from the ground up.
we have to figure out what constitutes death and game over. right now, the game is considered Over if a Player's final_health = 0. this is fine for single player, but doesn't make much sense for multiplayer, as the still-alive player should be able to continue, or even perhaps revive the dead player. thus a new system will have to be designed and implemented. there will also be balance considerations between playing solo & multi. if in multiplayer players are able to revive one another, than playing multiplayer has a clear and obvious advantage over playing solo.
i was toying with the idea in my idea of having multiple characters from whom we can choose to play as. if you die in single player, to retrieve the dead character, you must play as a different character up to the point of retrieving your corpse. this allows for a kind of "continue" system where death does not instantly mean game over, and also facilitates experimenting with different builds/stats/formulas on different characters. if all characters die, however, then the game is over and a save will need to be loaded. this gives the player up to 4 chances/deaths in a row, without any corpse retrieval, before hitting the game over screen. even if the player focuses on leveling up only 1 character, this system still seems fairly lenient and not overly punishing.
for multiplayer, it might look completely different. brainstorming, there is the "arcade" beat'em up way with shared continues. if a player dies, a continue must be used to revive. if both players use up all their continues, combined, the game is over. eg. one player has to die 4 times or both player 2 times each.
i like that idea, but the problem is that it doesn't lend well to drop-in/drop-out multiplayer. perhaps the final game will have several modes available, including an "arcade"-style mode that allows for drop-in/out multiplayer - but the story campaign must either be played fully solo or fully co-op. this would also make designing the story potentially simpler. there could be a series of narrative overlaps depending on characters chosen, but the game's overall narrative won't need to take into account at any point whether more than one player is active or not.
there can be a "solo campaign", which follows how a group of tenuously connected characters might traverse the story. how each character reacts to events will change depending on the character(s) present, but the overall flow from start to finish will be the same each playthrough. characters can be switched at "save points" or "hubs", the nature of which has not yet been decided.
then, there can be a "co-op campaign", which must be played by 2 players minimum. it could potentially work in a similar manner, with each player choosing a character, allowing for swaps at save points or hubs. or, specifically for co-op, player character choices are locked-in and cannot be changed for the duration of the playthrough. these are things to consider.
then, there could also be an "arcade" mode. arcade mode will have less emphasis on story cutscenes and narrative structure and more emphasis on getting through the game, and it is in THIS mode where we allow drop-in and out multiplayer. to emphasize gameplay, perhaps instead of character switching/corpse retrieval on death it works like an arcade beat'em up with limited continues for a single playthrough. i like the idea of having multiple modes to play through the game. just need to think about how to design the systems that surround those modes, and to make sure those systems are complimentary.