minor fixes
pressing the cancel button twice while on the pause screen no longer plays the menu closing animation twice. updated the distance the two players can move apart from each other.
now i have to figure out a better way to handle camera movement in multiplayer. before, the Player scripts were discrete, so i could hard-code how both player 1 & 2 affected the camera. now, players share a single script. there are two places in this script that affect the camera during multiplayer:
1. line 269 (move_state(delta), if input_vector != ZERO)
checks if player 2's distance from player 1 exceeds certain horizontal or vertical thresholds, and if it does, sets player 2's position to be that of player 1's (basically, teleports player2 to player1 if they're far away from each other)
2. line 374 (move() function)
as above, checks to see if the difference between player 1 and player 2's positions exceed certain thresholds. if a check succeeds and a limit is broken, begins moving player2 one pixel per frame on the axis associated with the broken threshold.

so the problem is that only player 2 ever gets teleported or moved. teleportation only occurs on player1 INPUT, so player2 can continue to move offscreen if there is no player1 input.

i would like to design a system that can potentially handle more than 2 players at a time. one of the issues with camera placement in multiplayer is if there is a conflict, who should teleport to the other player, 1 or 2? i believe the idea was that if one player was detected to have input and the other not, we teleport the player WITHOUT input. but what if both players have input, and are walking in opposite directions? it may make sense to simply give player 1 camera priority and have player 1 "win" these conflicts if they occur.
another potential issue, if we are to deal with more than two players, is who to teleport to. the default could be player1, or maybe it could be the closest player.