dying/leveling/zoning systems re-integrated
lots of small changes to get these fixes, let's see if i can remember most of them.
when we level, the levelTimer starts (0.4s) and the exits are disabled. when this levelTimer times out, and neither player is dying, the level up screen is instantiated. if only one of the players leveled up and no players are dying, the exits are re-enabled. if both players leveled up, the exits are only re-enabled after the final level up screen is instanced. if a player becomes "dying" while their levelTimer is running, the levelTimer is stopped - but their stats.leveling variable remains true.
if a player returns from dying and their stats.leveling = true, then we start their levelTimer UNLESS another player is dying.
basically, if a player recovers from dying, we check if any other players are also dying, and if they are, the dying effect remains enabled and the exits disabled. if a player levels while ANY player is dying, we set their stats.leveling to true, but don't start their levelTimer until all players are recovered from dying.
exits are disabled when a levelTimer starts, OR when a player is dying.
exits are re-enabled after all players recover from dying, and all level up screens are instanced.