fixes
ran into an issue where player HP was not properly set on Continue. the reason why is because max_health gets set AFTER health. meaning, if max_health is at its default value, whatever that is, the health value cannot exceed it. workaround was to reset default vitality to a high value upon entering the title screen, but also to set it to the default value of 4 when selecting new game.
also implemented ingredient examination persistence. meaning, the game save remembers which ingredients have been examined.
also made it so that when you press the examine button on something you have not yet examined (the "?!" notice appears), the ?! notice disappears immediately.
the only issue - and it isn't huge, but i notice it, so i'll fix it eventually - is that when both players are looking at the same unexamined object, and one player examines it, the other's ?! notice does not disappear until he either presses his own examine input, or the interactObject leaves his interactHitbox.
i'm not sure what would be an efficient way to solve this. perhaps on running each object's "examine_complete" method, we reset each player's interactHitbox, but that seems like a little much.