tuning interactions
when an object becomes fully examined, all players' interaction is reset. meaning, if both players are looking at an object and one of them examines it fully, the ?! notice will disappear for the other player after closing the dialog box.
there is some finnickyness here, but it's working, for the mostpart. the only thing right now that needs fixing for local multiplayer is inputs while a dialog box is open. right now, both players' inputs advance the dialog in dialog boxes. this can cause some weirdness, for example when closing a dialogbox that was opened by another player: the player closing the dialog box does not have their talkTimer started, therefore if they close the dialog box with their "examine" button, a new examine dialog box will appear instantly.
a potential solution is to have all players' talkTimers run when one runs. hell, if that's the case, may as well make the talkTimer global, ie. shared between all players. that actually sounds like a reasonably elegant solution. doesn't fix the issue of both players being able to advance dialog, but that may be a non-issue ultimately. i still need to implement proper player input for the level screen, however, and also to include an indicator as to WHO has leveled because it isn't exactly clear.