ui input
unfortunately godot is designed so that modifier keys like shift aren't to be used as regular keys. therefore using SHIFT+ something else returns a discrete input, which is weird and sucks, because i don't want SHIFT+ui_accept to be ui_accept, i want it to be nothing.
it sucks because if i code an input function to ignore a particular key, pressing SHIFT+ that key propogates the input anyway. that's what makes no sense to me. for illustration, say i have V in the InputMap as ui_accept. then, for a dialogbox that player2 instantiates, i have that dialog box's input method to set event as text V to be handled & ignored. if SHIFT is held, the V input is not ignored! this wouldn't be an issue except that the InputMap still recognized SHIFT+V as ui_accept!! despite SHIFT+V being an event that's discrete from just V!!!!
if shift+V exists as a discrete input event, why does it ALSO activate actions assigned to JUST V??? and of course there is no way to change this.