experimenting with outlines
ingredient pickups are outlined in cyan when inside a player's interactHitbox.
i used a shader material to achieve this effect, but it is not pixel perfect, so i may decide against it. another approach would be to simply create outlines for each object manually, and show/hide them as necessary. i have done this with the enemies, and i think it looks good. having the outline as its own Node also allows its Z-index to be manually set, so that outlines can display in front of obstacles - so you'd be able to see enemies/items obstructed by trees, for example.
the disadvantage to doing it this way is the annoyance and tedium. luckily aseprite has an easy-to-use outline tool. but if a sprite is drawn all the way to the border of a frame, then its frame size & spritesheet need to be updated accordingly.
right now there is a very hacky way of doing the outlines for enemies. basically just setting the frame of the outline to whatever frame the sprite is on every physics frame. there is probably a more efficient way to do this, like changing the outline frame only when the sprite's frame changes. i will figure out how to do that. unfortunately, the signal method didn't work the first time i tried it, i'm guessing because when the Sprite gets instantiated it shoots its signal, but the Outline hasn't been instantiated yet so it doesn't know what "frame" is.
but before i do that i may want to continue with the Enemy refactor. right now, enemies are not extended from the Enemy script: the Enemy script simply acts as a global list of functions common between enemies. i think it would make more sense of enemies extended this script. i will look into doing this.