fixes & adjustments
sword knockback_vector is now multiplied by the player's knockback_modifier. default=1, and during frenzy=0.2.
charge attacks override this. lv.1 sword charge knockback_vector is still multiplied by 1.33, and lv.2 sword charge by 0.
also fixed some issues where the game would throw an error/freeze if the scene was changed (like the game being quit) while a function was yielding, for example if you quit the game after killing an enemy but BEFORE it explodes, there would be an error.
i would like to sort out the message display for the two players. it seemed impossible to set the alignment via code, so i might have to create separate scenes for messages that display on the left or right side of the screen, which is stupid as fuck. or i can compromise and display all messages in the center.