created Slow status / other additions
not fully implemented. no slow for enemies yet, and lacking other effects & testing.
slow halves movement speed and the speed stat (which affects evasion & attack speed among other things). also adjusts the run TimeScale to be 0.5, or 50%, slowing the animation. i still have to insert TimeScale nodes for the other animations, which will be tedious.
crows attacks have a 50% chance to Slow. you can test slow on the players by pressing '3'.
since this is the first status that directly affects the player's stats, i have to be careful as to not permanently alter stats unintentionally. for example, before, hitting the save point would save the PlayerStats before reverting them from the slow status, resulting in a permanently halved speed stat. what i should do is create a rigid stat modifier system, which will always keep track of the player's base stats as well as those being buffed/debuffed. this is also necessary if i am to display stat modifiers in the menu.
right now, when the Slow status is applied to a player, it keeps track of the old speed stat inside of a variable: old_speed. when the status Timer times out - OR the status is rid some other way - old_speed is applied to player.stats.speed.