animated Flash formula target screen
giving it a little bit of dynamicism. tween scales the target reticle as well as the fade.
if a formula is cast, the game resumes instantly. if the target screen is cancelled out, the game unpauses only after 0.25s. though i may keep it paused for the duration either way, we'll see what feels better once more targettable spells are in.
on that note, i think i'll make all formulas targettable, even the large AoE ones like Heal and Fury - which, basically, just apply the buffs to everyone on screen. some spells won't have a movable reticle - the target screen will exist simply as a way to either confirm or cancel out of a cast.