fixed formula target screen next/previous issue
fixed issue in previous commit where pressing next/previous on the formula target screen would unintentionally alter the origin point to determine the formula range limit.
i have encountered another issue. if the save point is entered and the player presses the formula button at the same time, the formula target screen is instantiated AND the save function is run. the issue is the save function unpauses the game when it's finished, even if you are still on the formula target screen. you can still either cancel out or select where to cast, only the game is unpaused while this is happening - which, not gonna lie, is potentially cool. i think i would like to include a mode where there is no pause while targetting. kind of like how old FF games had ACTIVE or WAIT mode.
anyway solution might be to have the various aspects of the game that pause/unpause it also toggle global variables, which will also be checked when pausing/unpausing. eg., the formula target screen declares global variable "on_target_screen" to be true. whenever we attempt to unpause the game, we check these variables, and don't unpause until all of them are false.