fixing minor target screen issues
fixed an issue where the enemy targets weren't becoming red when overlapping. the problem was due to where we pause the game. if we pause too early, like when the Flash scene is instanced, then the targets' collision does not get detected. pausing when the target screen is instanced would allow for the previous commit's bug to occur, which is that when a formula is cast while walking into a save point, the completion of the save function would un-pause the game, even if the target screen was open.
the solution for now is to declare a Global variable, target_screen_open, and not unpause the game if it's true. the variable becomes true when a formula is cast, and becomes false when the target screen is ended.
i will have to be very careful moving forward with events that pause/unpause the game via timing, like the save point. the save point will likely have other functions that will be available from a menu, so it won't exist as it currently does, but it'll be important to remember.