PlayerStats overhaul nearing completion
stat modifiers are coded and functioning properly.
previous stat modifiers have been renamed and refactored (eg. old dexterity_mod is now dexterity_bonus, old magic_mod is now spell_mod, old speed_mod is now max_speed_mod, etc.) to make room for the new stat modifier variables, which affect the range of secondary stats that are governed by the primary stats.
it is not yet finished and far from fully implemented, but the foundation is nearing completion.
an example for illustration...
the primary stat "Patience", aka stats.dexterity, now governs a secondary stat called accuracy (stats.accuracy). when an enemy is attacked, we subtract the enemies "evasion" from our "accuracy", and run the result against the RNG.
when we change our dexterity_mod (the number by which we multiply stats.dexterity to get our stats.accuracy value), our stats.accuracy automatically gets updated. the same occurs when we change our dexterity_bonus (the number by which we increment stats.dexterity) - stats.accuracy automatically gets calculated, and it is this stats.accuracy that is checked when we attack an enemy.
next i will refactor the crit calculation. instead of calculating crit chance every time we hit an enemy, it will be calculated automatically when we set our dexterity, and it will be from this value against which we check the RNG for crits.
now that i think of it, once this system overhaul is complete, i'm not sure it will be necessary for players' and enemies' hit/crit calculation to be different functions. since the values are going to be precalculated, when we run the RNG for hits/crits, we are simply checking input values (accuracy-evasion, crit chance) against RNG, and returning either true or false. a simple function like this should be universally usable by both players and enemies, further simplifying the code.