crit calculation refactor
players & enemies now share the same crit check function.
the new crit check function takes a single argument, crit_rate, which is then checked against RNG, and then either returns true or false. simple as can be.
player crit_rate now takes place in the PlayerStats scene. currently, it is:
base_crit_rate + (dexterity * dexterity_mod + dexterity_bonus)/4
where base_crit_rate = 4, and dexterity_bonus for a Lv.2 charge attack is 8. let's say our player has 12 dex and no dexterity_mod (1). the final crit chance using the above parameters would be 9%:
4 + (12*1+8) / 4 = 9.
next i want to fix how base_enemy_accuracy is handled. right now, it is hard-coded on the Player with a default value of 66. various player actions change this value: rolling makes it 32, and backstepping makes it 16; a Lv.1 charge attack makes it 25, and a Lv1. charge attack queued during a backstep makes it 0. basically, the player is harder to hit during certain actions.
so the base_enemy_accuracy is 66, and each enemy has their own stats.accuracy. enemy hit calculation is currently handled as such:
base_enemy_accuracy + (4*accuracy) - evasion = final_hit_rate.
this is kind of a weird formula. it might make more intuitive sense to simple get rid of the base_enemy_accuracy stat altogether and instead just give each individual enemy their own accuracy stat. the reason i originally coded it this way was because i'd figured enemy accuracy would remain pretty consistent throughout the entire game, so the individual enemies' accuracy stats basically served as an accuracy bonus on top of the hard-coded base_enemy_accuracy of 66. but now that we are working on a much more rigid stat system, the calculation can be done in a more typical RPG fashion. the trickiness will be how to factor in the changes to base_enemy_accuracy due to the various player actions that change it. maybe i will do it the same way dexterity_bonus is applied, except it will be an evasion_bonus instead. that sounds fine so i'll be going with that.